backend/requirements.txt
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.2.0
pandas==2.2.3
psycopg==3.2.6
psycopg-binary==3.2.6
